Cleaned up a little.
