+++ src/plugins/skills/smithing/smelting-types.ts
@@ -1,0 +1,15 @@
+import { Item } from '@server/world/items/item';
+
+export interface Bar {
+    barId: number;
+    quest?: string;
+    requiredLevel: number;
+    ingredients: Item[];
+    experience: number;
+}
+
+export interface Smeltable {
+    takesInput: boolean;
+    count: number;
+    bar: Bar;
+}
